Vite example app
