+++ example/index.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="no">
+    <head>
+        <meta charset="utf-8" />
+        <meta
+            name="viewport"
+            content="width=device-width, initial-scale=1, shrink-to-fit=no"
+        />
+        <meta
+            http-equiv="Cache-control"
+            content="no-cache, no-store, must-revalidate"
+        />
+        <meta http-equiv="Pragma" content="no-cache" />
+        <meta name="theme-color" content="#000000" />
+        <link rel="manifest" href="/manifest.json" />
+        <link rel="shortcut icon" href="/favicon.ico" />
+        <link rel="stylesheet" href="/index.css" />
+        {{{STYLES}}}
+        <script
+            type="application/javascript"
+            src="/frontendlogger/logger.js"
+        ></script>
+        <title>nav-dekoratoren-moduler</title>
+    </head>
+    <body>
+        <noscript> You need to enable JavaScript to run this app. </noscript>
+        <div class="pagewrapper">
+            {{{HEADER}}}
+            <div class="app" id="app"></div>
+        </div>
+        {{{FOOTER}}} {{{SCRIPTS}}}
+        <script type="module" src="/src/main.tsx"></script>
+    </body>
+</html>
