+++ example/package.json
@@ -1,43 +1,29 @@
 {
-    "name": "@navikt/nav-dekoratoren-moduler-example",
-    "homepage": ".",
-    "version": "0.0.0",
+    "name": "example-ts",
     "private": true,
+    "version": "0.0.0",
+    "type": "module",
     "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build"
+        "dev": "vite",
+        "build": "tsc && vite build",
+        "lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+        "preview": "vite preview"
     },
     "dependencies": {
         "@navikt/nav-dekoratoren-moduler": "file:..",
-        "jsdom": "^16.6.0",
-        "node-cache": "^5.1.2",
-        "node-fetch": "^2.6.1",
-        "react": "file:../node_modules/react",
-        "react-dom": "file:../node_modules/react-dom",
-        "react-scripts": "file:../node_modules/react-scripts",
-        "typescript": "file:../node_modules/typescript"
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0"
     },
     "devDependencies": {
-        "@types/jsdom": "^16.2.13",
-        "@types/node": "file:../node_modules/@types/node",
-        "@types/node-fetch": "^2.5.11",
-        "@types/react": "file:../node_modules/@types/react",
-        "@types/react-dom": "file:../node_modules/@types/react-dom",
-        "@types/react-html-parser": "^2.0.1"
-    },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
+        "@types/react": "^18.0.37",
+        "@types/react-dom": "^18.0.11",
+        "@typescript-eslint/eslint-plugin": "^5.59.0",
+        "@typescript-eslint/parser": "^5.59.0",
+        "@vitejs/plugin-react": "^4.0.0",
+        "eslint": "^8.38.0",
+        "eslint-plugin-react-hooks": "^4.6.0",
+        "eslint-plugin-react-refresh": "^0.3.4",
+        "typescript": "file:../node_modules/typescript",
+        "vite": "^4.3.9"
     }
 }
