--- example/public/index.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="no">
-<head>
-  <meta charset="utf-8" />
-  <meta
-    name="viewport"
-    content="width=device-width, initial-scale=1, shrink-to-fit=no"
-  />
-  <meta
-    http-equiv="Cache-control"
-    content="no-cache, no-store, must-revalidate"
-  />
-  <meta http-equiv="Pragma" content="no-cache" />
-  <meta name="theme-color" content="#000000" />
-  <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-  <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
-  <link rel="stylesheet" href="%PUBLIC_URL%/index.css" />
-  {{{STYLES}}}
-  <script
-    type="application/javascript"
-    src="/frontendlogger/logger.js"
-  ></script>
-  <title>nav-dekoratoren-moduler</title>
-</head>
-<body>
-<noscript> You need to enable JavaScript to run this app. </noscript>
-<div class="pagewrapper">
-  {{{HEADER}}}
-  <div class="app" id="app"></div>
-</div>
-{{{FOOTER}}} {{{SCRIPTS}}}
-</body>
-</html>
