+++ example/public/manifest.json
@@ -1,15 +1,15 @@
 {
-  "short_name": "@navikt/nav-dekoratoren-moduler",
-  "name": "@navikt/nav-dekoratoren-moduler",
-  "icons": [
-    {
-      "src": "favicon.ico",
-      "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/x-icon"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
+    "short_name": "@navikt/nav-dekoratoren-moduler",
+    "name": "@navikt/nav-dekoratoren-moduler",
+    "icons": [
+        {
+            "src": "favicon.ico",
+            "sizes": "64x64 32x32 24x24 16x16",
+            "type": "image/x-icon"
+        }
+    ],
+    "start_url": ".",
+    "display": "standalone",
+    "theme_color": "#000000",
+    "background_color": "#ffffff"
 }
