+++ example/src/index.css
@@ -1,41 +1,40 @@
 body {
-  margin: 0;
-  padding: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+    margin: 0;
+    padding: 0;
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto",
+        "Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans",
+        "Helvetica Neue", sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+    font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
+        monospace;
 }
 
-.decorator-header{
-  min-height:150px
+.decorator-header {
+    min-height: 150px;
 }
 
-.decorator-footer{
-  min-height: 783px
+.decorator-footer {
+    min-height: 783px;
 }
 
-
-.testapp__textarea{
-  height: 250px;
-  width:500px;
+.testapp__textarea {
+    height: 250px;
+    width: 500px;
 }
 
-.testapp__container{
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-  height:35rem;
-  width:100%
+.testapp__container {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    align-items: center;
+    height: 35rem;
+    width: 100%;
 }
 
-.testapp__container > *{
-  margin: 0.5rem
+.testapp__container > * {
+    margin: 0.5rem;
 }
